pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-registry-scraper"
version = "1.1.0"
description = "A comprehensive tool for scraping the entire MCP server registry to collect metadata, tool descriptions, and server information with enhanced multi-language support"
authors = [
    {name = "MCP Registry Scraper Contributors"}
]
keywords = ["mcp", "model-context-protocol", "scraper", "github", "neo4j", "tools", "ecosystem-analysis"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "PyGithub>=1.59.1",
    "pydantic>=2.5.0",
    "python-dateutil>=2.8.2",
    "tqdm>=4.66.1",
    "aiohttp>=3.9.1",
    "asyncio-throttle>=1.0.2",
    "neo4j>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/your-username/github-mcp-scraper"
"Bug Reports" = "https://github.com/your-username/github-mcp-scraper/issues"
"Source" = "https://github.com/your-username/github-mcp-scraper"
Documentation = "https://github.com/your-username/github-mcp-scraper#readme"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "isort>=5.10.0",
]
neo4j = [
    "neo4j>=5.0.0",
]
visualization = [
    "matplotlib>=3.5.0",
    "networkx>=2.8.0",
]

[project.scripts]
mcp-scraper = "mcp_scraper.main:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["mcp_scraper*"]
exclude = ["tests*"]